Move custom derive test to external test file
