+++ fitsio-derive/src/lib.rs
@@ -34,7 +34,10 @@ fn impl_read_row(input: syn::DeriveInput) -> quote::Tokens {
 
     quote!{
         impl FitsRow for #name {
-            fn from_table(tbl: &FitsHdu, fits_file: &mut FitsFile, idx: usize) -> Result<Self> where Self: Sized  {
+            fn from_table(
+                tbl: &::fitsio::FitsHdu,
+                fits_file: &mut
+                FitsFile, idx: usize) -> ::fitsio::errors::Result<Self> where Self: Sized  {
                 fits_file.make_current(tbl)?;
                 let mut out = Self::default();
